refactor: Replace is_local_search boolean with SearchMode enum + add eviction tests (#347)
* refactor: Replace is_local_search boolean with SearchMode enum and add eviction tests

**API Refactoring:**
- Add SearchMode enum with kLocal and kRemote values for better code clarity
- Update Search() and SearchAsync() function signatures to use SearchMode
- Replace all boolean calls with appropriate enum values:
  - true -> SearchMode::kLocal (for direct local searches)
  - false -> SearchMode::kRemote (for distributed/coordinator searches)
- Improves code readability and self-documentation

**Testing Enhancement:**
- Add comprehensive eviction test suite (integration/test_eviction.py)
- Tests search index behavior with all Valkey eviction policies
- Validates proper cleanup of evicted hashes from search indexes
- Ensures search functionality remains intact after eviction events

No functional changes to search behavior, only improved API design and test coverage.

+++ vmsdk/src/debug.cc
@@ -119,7 +119,7 @@ void PausePointList(ValkeyModuleCtx* ctx) {
       ValkeyModule_ReplyWithSimpleString(ctx, "Location");
       ValkeyModule_ReplyWithSimpleString(ctx, ToString(w.location_).data());
       ValkeyModule_ReplyWithSimpleString(ctx, "Threadid");
-      ValkeyModule_ReplyWithLongLong(ctx, w.threadid_);
+      ValkeyModule_ReplyWithLongLong(ctx, (long long)(uintptr_t)w.threadid_);
       ValkeyModule_ReplyWithSimpleString(ctx, "WaitSeconds");
       ValkeyModule_ReplyWithDouble(
           ctx, absl::ToDoubleSeconds(absl::Now() - w.start_time_));
